requirements.txt
Django==4.2.24
djangorestframework==3.16.1
cryptography==45.0.7
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2
typing-extensions==4.15.0
django-cors-headers==4.3.1

